add basic array slicing implementation in ndview

+++ src/serac/numerics/functional/refactor/refactor_interpolation_test.cpp
@@ -26,7 +26,7 @@ int main(int argc, char* argv[]) {
   u_T = 1.0;
   axom::Array < double, 2 > u_Q;
   axom::Array < double, 3 > du_dxi_Q;
-  interpolate(u_Q, du_dxi_Q, u_T, q);
+  interpolate(u_Q, du_dxi_Q, u_T, uint32_t(q));
 
   //for (int i = 0; i < u_Q.Dims[0]; i++) {
   //  for (int j = 0; j < u_Q.Dims[1]; j++) {
